Added new target to obtain compiler pid.
One of things that my vim does when saving (writing) LaTeX files, is to
automagically re-compile them. To avoid clashes if there is already a
compilation process running, vim won't interrupt it. In order to detect
whether a job is running or not, this new makefile phony target is
provided.