Updates are mainly to setup.sh and CompileTeX.sh, the rest is small stuff (like updating my domain name...).
First, on gitignore, replace Makefile (which was deleted) shell script
(and sort alphabetically).

Next, for CompileTeX.sh, added the stuff to make a final copy, an
replaced `cd -` with `cd ..`. Because in some cases, cd - was run
twice, and so the program remained in the same dir.

And on setup.sh, the variables in sed commands were updated, to
reflect the new names. Instructions were added to deal with
unabridged folder. Also now exists, in CompileTeX.sh, a boolean
to indicate whether bibliography is being used or not; depending
on the doc type, this is also sed'd accordingly.

In report, a fancy header, that displays page number and chapter
name and number, on top, was added.