Unabridged copy, for reports, is now built in a new way...
Instead of copying over all the .tex files into a new directory, and
then constructing a new unabridged copy therein, we now instead copy
only the main .tex file, to a new file named Unabridged.tex also in
the root folder, comment any \includeonly lines, if any, and then build
the unabridged copy in a new build dir, viz build_UNABRIDGED.

This required some changes, namely to the run method, which now
takes as arguments the .tex filename and the build dir.

But overall, this is a far simpler strategy.