fullrun is now big_compile...
fullrun might suggest we build the "full" document, which is not the
case, if using eg, \includeonly. So it is best to say we are doing a
"big compile": latex (and index if applicable), bibtex, latex, latex,
latex.

Also, added new error messages, even for unabridged copy compile and
bibtex.