The $got_bib var is now called $do_bib.
This point is we leave the bib stuff in place always, and check the
existence of \cite commans to see whether or not we build it.

The var is just to allow the user to explicitly disable the bib, in case
of some error he does want to deal with right now, or something like
that.