Compile script: two new functions
Compile script now has an option rebuild_aux to rebuild the .aux files
database used by \include[only].

Also new function u2r, which just copies the unabridged copy with the
name of the regular copy. Thus we are left with two complete copies, to
ease reading. THis is useful when one intends to actually read the doc,
with only the occasional writing.